requirements.txt
# Django Core
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Database
psycopg2-binary==2.9.9

# AI Integration
openai==1.12.0
google-generativeai==0.4.0

# Environment Variables
python-decouple==3.8

# Caching
redis==5.0.1
django-redis==5.4.0

# Static Files (Production)
whitenoise==6.6.0

# Image Processing
Pillow==10.2.0

# UUID Support
uuid==1.30

# Date/Time
python-dateutil==2.8.2

# Astronomy & Astrology
pyswisseph==2.10.3.2

# Geocoding
geopy==2.4.1

# Production Server
gunicorn==21.2.0

# Development Tools
django-debug-toolbar==4.2.0

# Security
django-cors-headers==4.3.1

# Forms
django-widget-tweaks==1.5.0